+++ src/Data/Equality/Analysis.hs
@@ -25,7 +25,7 @@ module Data.Equality.Analysis where
 import Data.Equality.Graph.Classes.Id
 import Data.Equality.Graph.Nodes
 
-import {-# SOURCE #-} Data.Equality.Graph (EGraph)
+import {-# SOURCE #-} Data.Equality.Graph.Internal (EGraph)
 
 -- | The e-class analysis defined for a language @l@.
 class Eq (Domain l) => Analysis l where