+++ src/Data/Equality/Graph.hs
@@ -3,7 +3,6 @@
 {-# LANGUAGE TupleSections #-}
 -- {-# LANGUAGE ApplicativeDo #-}
 {-# LANGUAGE BlockArguments #-}
-{-# LANGUAGE UndecidableInstances #-} -- tmp show
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TypeFamilies #-}
@@ -15,9 +14,7 @@
 module Data.Equality.Graph
     (
       -- * Definition of e-graph
-      EGraph(..)
-
-    , Memo, Worklist
+      EGraph
 
       -- * Functions on e-graphs
     , emptyEGraph
@@ -39,11 +36,10 @@ module Data.Equality.Graph
 
 import Data.Function
 
-import Data.Functor.Classes
-
 import qualified Data.IntMap.Strict as IM
 import qualified Data.Set    as S
 
+import Data.Equality.Graph.Internal
 import Data.Equality.Graph.ReprUnionFind
 import Data.Equality.Graph.Classes
 import Data.Equality.Graph.Nodes
@@ -51,38 +47,11 @@ import Data.Equality.Analysis
 import Data.Equality.Language
 import Data.Equality.Graph.Lens
 
--- | E-graph representing terms of language @l@.
---
--- Intuitively, an e-graph is a set of equivalence classes (e-classes). Each e-class is a
--- set of e-nodes representing equivalent terms from a given language, and an e-node is a function
--- symbol paired with a list of children e-classes.
-data EGraph l = EGraph
-    { unionFind :: !ReprUnionFind           -- ^ Union find like structure to find canonical representation of an e-class id
-    , classes   :: !(ClassIdMap (EClass l)) -- ^ Map canonical e-class ids to their e-classes
-    , memo      :: !(Memo l)                -- ^ Hashcons maps all canonical e-nodes to their e-class ids
-    , worklist  :: !(Worklist l)            -- ^ Worklist of e-class ids that need to be upward merged
-    , analysisWorklist :: !(Worklist l)     -- ^ Like 'worklist' but for analysis repairing
-    }
-
--- | The hashcons 𝐻  is a map from e-nodes to e-class ids
-type Memo l = NodeMap l ClassId
-
--- | Maintained worklist of e-class ids that need to be “upward merged”
-type Worklist l = NodeMap l ClassId
-
 -- ROMES:TODO: join things built in paralell?
 -- instance Ord1 l => Semigroup (EGraph l) where
 --     (<>) eg1 eg2 = undefined -- not so easy
 -- instance Ord1 l => Monoid (EGraph l) where
 --     mempty = EGraph emptyUF mempty mempty mempty
-
-instance (Show (Domain l), Show1 l) => Show (EGraph l) where
-    show (EGraph a b c d e) =
-        "UnionFind: " <> show a <>
-            "\n\nE-Classes: " <> show b <>
-                "\n\nHashcons: " <> show c <>
-                    "\n\nWorklist: " <> show d <>
-                        "\n\nAnalWorklist: " <> show e
 
 
 -- | Add an e-node to the e-graph
@@ -234,7 +203,7 @@ merge a b egr0 =
 {-# SCC merge #-}
             
 
--- | The rebuild operation processes the e-graph's current 'Worklist',
+-- | The rebuild operation processes the e-graph's current worklist,
 -- restoring the invariants of deduplication and congruence. Rebuilding is
 -- similar to other approaches in how it restores congruence; but it uniquely
 -- allows the client to choose when to restore invariants in the context of a